DEV-1427 Run pipeline within the quotas of a project
We wrap the compute engine with a decorator which checks the max cpu quota and
scales down the CPU ceiling of any VM to ensure we don't try to spawn a VM
which will always fail.

Then every job submit is wrapped with a retry, so if the issue is that other jobs
are taking all the quota, the idea is we'll just wait till more room opens up.
