+++ batch/src/test/java/com/hartwig/batch/BatchDispatcherTest.java
@@ -20,9 +20,10 @@ import com.hartwig.batch.input.InputParser;
 import com.hartwig.batch.input.InputParserProvider;
 import com.hartwig.batch.operations.OperationDescriptor;
 import com.hartwig.batch.testsupport.TestingArguments;
-import com.hartwig.pipeline.execution.vm.ComputeEngine;
+import com.hartwig.pipeline.execution.vm.GoogleComputeEngine;
 
 import org.junit.Test;
+import org.mockito.Mockito;
 
 public class BatchDispatcherTest {
     @Test
@@ -36,7 +37,7 @@ public class BatchDispatcherTest {
         InputParserProvider inputParserProvider = mock(InputParserProvider.class);
         InputParser inputParser = mock(InputParser.class);
         BatchOperation operation = mock(BatchOperation.class);
-        ComputeEngine computeEngine = mock(ComputeEngine.class);
+        GoogleComputeEngine computeEngine = mock(GoogleComputeEngine.class);
         Storage storage = mock(Storage.class);
         Bucket bucket = mock(Bucket.class);
         ExecutorService executorService = mock(ExecutorService.class);
@@ -48,7 +49,7 @@ public class BatchDispatcherTest {
         when(inputParser.parse(any(), any())).thenReturn(asList(bundle1, bundle2));
         when(bundle1.get()).thenReturn(mock(InputFileDescriptor.class));
         when(bundle2.get()).thenReturn(mock(InputFileDescriptor.class));
-        when(executorService.submit((Callable) any())).thenReturn(mock(Future.class)).thenReturn(mock(Future.class));
+        when(executorService.submit(Mockito.<Callable<?>>any())).thenReturn(mock(Future.class)).thenReturn(mock(Future.class));
 
         when(storage.get(arguments.outputBucket())).thenReturn(bucket);
 
@@ -56,6 +57,6 @@ public class BatchDispatcherTest {
                 computeEngine, storage, executorService);
 
         dispatcher.runBatch();
-        verify(executorService, times(2)).submit((Callable) any());
+        verify(executorService, times(2)).submit(Mockito.<Callable<?>>any());
     }
 }
