+++ pom.xml
@@ -41,6 +41,7 @@
         <commons-lang3.version>3.9</commons-lang3.version>
         <htsjdk.version>2.22.0</htsjdk.version>
         <org-reflections.version>0.9.11</org-reflections.version>
+        <google-api-services-serviceusage.version>v1beta1-rev20200804-1.30.10</google-api-services-serviceusage.version>
     </properties>
 
     <scm>
@@ -160,6 +161,11 @@
                 <groupId>com.google.apis</groupId>
                 <artifactId>google-api-services-cloudbilling</artifactId>
                 <version>${google-billing.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.google.apis</groupId>
+                <artifactId>google-api-services-serviceusage</artifactId>
+                <version>${google-api-services-serviceusage.version}</version>
             </dependency>
             <dependency>
                 <groupId>com.google.protobuf</groupId>
